:heavy_minus_sign: remove drf-yasg
