+++ src/objects/conf/base.py
@@ -64,7 +64,6 @@ INSTALLED_APPS = [
     # 'django.contrib.sitemaps',
     # External applications.
     "axes",
-    "drf_yasg",
     "drf_spectacular",
     "sniplates",
     "hijack",
